riboviria/requirements.txt
certifi==2019.6.16
chardet==3.0.4
click==7.1.1
Flask==1.1.2
Flask-SQLAlchemy==2.4.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
paho-mqtt==1.4.0
requests==2.22.0
SQLAlchemy==1.3.16
urllib3==1.25.3
Werkzeug==1.0.1
